late

+++ src/main/java/com/bootapp/rest/restapp/controller/CategoryController.java
@@ -2,9 +2,11 @@ package com.bootapp.rest.restapp.controller;
 
 import java.util.List;
 import java.util.Optional;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,49 +15,71 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+
+import com.bootapp.rest.restapp.dto.Message;
 import com.bootapp.rest.restapp.model.Category;
+import com.bootapp.rest.restapp.model.Customer;
 import com.bootapp.rest.restapp.service.CategoryService;
 import com.rest.restapp.Exception.CategoryNotFoundException;
+import com.rest.restapp.Exception.NullValueException;
 
 @RestController
+@CrossOrigin(origins = {"*"})
 @RequestMapping("/api/category")
 public class CategoryController {
+	
 	@Autowired
 	private CategoryService categoryservice;
-
-	@PostMapping("/add")
- public ResponseEntity<String> postCategory(@RequestBody Category category) throws Exception {
- categoryservice.insertBook(category);
- return ResponseEntity.status(HttpStatus.OK).body("Category posted in DB");
- }
 	
-
+	@PostMapping("/add")
+	public ResponseEntity<Message> insertCategory(@RequestBody Category category) {
+		Message m = new Message();
+		try {
+			categoryservice.insertCategory(category);
+			m.setMsg("Category added");
+			return ResponseEntity.status(HttpStatus.OK).body(m);
+		}
+		catch(Exception e) {
+			m.setMsg("Could not process the request, Try Again");
+			return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(m);
+		}
+			
+		}
+	
+//	@PostMapping("/add")
+//	public ResponseEntity<String> postCategory(@RequestBody Category category) throws NullValueException {
+//		categoryservice.insertBook(category);
+//		return ResponseEntity.status(HttpStatus.OK).body("Category posted in DB");
+//		
+//	}
 	@GetMapping("/getall")
- public List<Category> getAllCategories() {
- List<Category> list = categoryservice.getAllCategories();
- return list;
- }
-
+	public List<Category> getAllCategories() {
+		List<Category> list = categoryservice.getAllCategories();
+		return list;
+	}
+	
 	@PutMapping("/edit/{categoryId}")
- public ResponseEntity<String> editCategory(@PathVariable("categoryId") int categoryId,@RequestBody Category categoryNew) {
- /* Step 1: check if this id given is valid by fetching the record from DB */
- Optional<Category> optional = categoryservice.getCategoryById(categoryId);
- if(!optional.isPresent()) {
- return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("invalid ID");
- }
- Category categoryDB = optional.get();
- if(categoryNew.getCategoryName() != null)
-categoryDB.setCategoryName(categoryNew.getCategoryName());
-categoryservice.PostCategory(categoryDB);
- return ResponseEntity.status(HttpStatus.OK).body("category record Edited..");
- }
- 
+	 public ResponseEntity<String> editCategory(@PathVariable("categoryId") int categoryId,@RequestBody Category categoryNew) {
+	 /* Step 1: check if this id given is valid by fetching the record from DB */
+	Optional<Category> optional = categoryservice.getCategoryById(categoryId);
+	if(!optional.isPresent()) {
+	 return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("invalid ID");
+	}
+	Category categoryDB = optional.get(); 
+	 if(categoryNew.getName() != null)
+	 categoryDB.setName(categoryNew.getName());
+	 
+	categoryservice.PostCategory(categoryDB);
+	return ResponseEntity.status(HttpStatus.OK).body("category record Edited..");
+	}
+	
+	 @DeleteMapping("/delete/{categoryId}")
+	 public ResponseEntity<Object> deleteCategoryById(@PathVariable("categoryId") int categoryId) throws CategoryNotFoundException {
+	 Optional<Category> optional = categoryservice.getCategoryById(categoryId);
+	 if (!optional.isPresent())return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid ID Given");
+	 categoryservice.deleteCategoryById(categoryId);
+	 return ResponseEntity.status(HttpStatus.OK).body("Category is deleted");
+	 }
 
-@DeleteMapping("/delete/{categoryId}")
- public ResponseEntity<Object> deleteCategoryById(@PathVariable("categoryId") int categoryId) throws CategoryNotFoundException {
- // Optional<Category> optional = categoryservice.getCategoryById(categoryId);
- //if (!optional.isPresent())return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Invalid ID Given");
- categoryservice.deleteCategoryById(categoryId);
- return ResponseEntity.status(HttpStatus.OK).body("Category is deleted");
- }
+
 }
